don't remove someone else's lock file

+++ lib/lock_file.rb
@@ -17,18 +17,21 @@ module Syncophant
     end
     
     def lock!
-      File.open(lock_file, "a") { |file| file.write(Process.pid) }
-      File.open(lock_file, "r") { |file| raise LockFileExists unless file.gets.to_i == Process.pid }
+      File.open(lock_file, "a") { |file| file.write("#{Process.pid}\n") }
+      raise LockFileExists unless lock_file_acquired?
     end
     
     def unlock!
       begin
-        File.delete(lock_file)
+        File.delete(lock_file) if lock_file_acquired? 
       rescue
         raise LockFileMissing
       end
     end
 
+    def lock_file_acquired?
+      File.open(lock_file, "r") { |file| file.gets.to_i == Process.pid }
+    end
   end
   
   class LockFileExists < StandardError; end
